Fixes bug 1490734 - Fresh profile can sometimes display no card in the third card position
